9980: refactor integration, integration green
